+++ openstl/core/metrics.py
@@ -38,7 +38,8 @@ def PSNR(pred, true):
 
 
 def metric(pred, true, mean=None, std=None, metrics=['mae', 'mse'],
-           clip_range=[0, 1], spatial_norm=False):
+           clip_range=[0, 1], channel_names=None,
+           spatial_norm=False, return_log=True):
     """The evaluation function to output metrics.
 
     Args:
@@ -48,7 +49,10 @@ def metric(pred, true, mean=None, std=None, metrics=['mae', 'mse'],
         std (tensor): The std of the preprocessed video data.
         metric (str | list[str]): Metrics to be evaluated.
         clip_range (list): Range of prediction to prevent overflow.
+        channel_names (list | None): The name of different channels.
         spatial_norm (bool): Weather to normalize the metric by HxW.
+        return_log (bool): Whether to return the log string.
+
     Returns:
         dict: evaluation results
     """
@@ -61,15 +65,45 @@ def metric(pred, true, mean=None, std=None, metrics=['mae', 'mse'],
     invalid_metrics = set(metrics) - set(allowed_metrics)
     if len(invalid_metrics) != 0:
         raise ValueError(f'metric {invalid_metrics} is not supported.')
+    if isinstance(channel_names, list):
+        assert pred.shape[2] % len(channel_names) == 0 and len(channel_names) > 1
+        c_group = len(channel_names)
+        c_width = pred.shape[2] // c_group
+    else:
+        channel_names, c_group, c_width = None, None, None
 
     if 'mse' in metrics:
-        eval_res['mse'] = MSE(pred, true, spatial_norm)
+        if channel_names is None:
+            eval_res['mse'] = MSE(pred, true, spatial_norm)
+        else:
+            mse_sum = 0.
+            for i, c_name in enumerate(channel_names):
+                eval_res[f'mse_{str(c_name)}'] = MSE(pred[:, :, i*c_width: (i+1)*c_width, ...],
+                                                     true[:, :, i*c_width: (i+1)*c_width, ...], spatial_norm)
+                mse_sum += eval_res[f'mse_{str(c_name)}']
+            eval_res['mse'] = mse_sum / c_group
 
     if 'mae' in metrics:
-        eval_res['mae'] = MAE(pred, true, spatial_norm)
+        if channel_names is None:
+            eval_res['mae'] = MAE(pred, true, spatial_norm)
+        else:
+            mae_sum = 0.
+            for i, c_name in enumerate(channel_names):
+                eval_res[f'mae_{str(c_name)}'] = MAE(pred[:, :, i*c_width: (i+1)*c_width, ...],
+                                                     true[:, :, i*c_width: (i+1)*c_width, ...], spatial_norm)
+                mae_sum += eval_res[f'mae_{str(c_name)}']
+            eval_res['mae'] = mae_sum / c_group
 
     if 'rmse' in metrics:
-        eval_res['rmse'] = RMSE(pred, true, spatial_norm)
+        if channel_names is None:
+            eval_res['rmse'] = RMSE(pred, true, spatial_norm)
+        else:
+            rmse_sum = 0.
+            for i, c_name in enumerate(channel_names):
+                eval_res[f'rmse_{str(c_name)}'] = RMSE(pred[:, :, i*c_width: (i+1)*c_width, ...],
+                                                       true[:, :, i*c_width: (i+1)*c_width, ...], spatial_norm)
+                rmse_sum += eval_res[f'rmse_{str(c_name)}']
+            eval_res['rmse'] = rmse_sum / c_group
 
     pred = np.maximum(pred, clip_range[0])
     pred = np.minimum(pred, clip_range[1])
@@ -88,8 +122,9 @@ def metric(pred, true, mean=None, std=None, metrics=['mae', 'mse'],
                 psnr += PSNR(pred[b, f], true[b, f])
         eval_res['psnr'] = psnr / (pred.shape[0] * pred.shape[1])
 
-    for k, v in eval_res.items():
-        eval_str = f"{k}:{v}" if len(eval_log) == 0 else f", {k}:{v}"
-        eval_log += eval_str
+    if return_log:
+        for k, v in eval_res.items():
+            eval_str = f"{k}:{v}" if len(eval_log) == 0 else f", {k}:{v}"
+            eval_log += eval_str
 
     return eval_res, eval_log