+++ tools/prepare_data/download_weatherbench.sh
@@ -10,6 +10,8 @@ wget "https://dataserv.ub.tum.de/s/m1524895/download?path=%2F5.625deg%2F2m_tempe
 # down 1.40625deg `2m_temperature` (128x256) and place them in `data/weather/` according to `https://github.com/pangeo-data/WeatherBench`
 # wget "https://dataserv.ub.tum.de/s/m1524895/download?path=%2F1.40625deg%2F2m_temperature&files=2m_temperature_1.40625deg.zip" -O 2m_temperature_1.40625deg.zip
 
+# please download other datasets in weather bench according to examples above!
+
 echo "finished"
 
 
@@ -20,6 +22,10 @@ echo "finished"
 #     │   ├── 2m_temperature
 #     │   ├── 10m_u_component_of_wind
 #     │   ├── 10m_v_component_of_wind
+#     │   ├── geopotential
 #     │   ├── relative_humidity
+#     │   ├── temperature
 #     │   ├── total_cloud_cover
+#     │   ├── u_component_of_wind
+#     │   ├── v_component_of_wind
 #     │   ├── ...